#1 | Async API more generic
